refactor: rename bitcoin init to adonai
