+++ src/ipc/interfaces.cpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2021-2022 The Bitcoin Core developers
+// Copyright (c) 2014-2022 The Bitcoin Core developers
 // Modifications (c) 2025 The Adonai Core developers
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
@@ -66,10 +66,10 @@ public:
             // Treat "auto" the same as "unix" except don't treat it an as error
             // if the connection is not accepted. Just return null so the caller
             // can work offline without a connection, or spawn a new
-            // bitcoin-node process and connect to it.
+            // adonai-node process and connect to it.
             address = "unix";
             try {
-                fd = m_process->connect(gArgs.GetDataDirNet(), "bitcoin-node", address);
+                fd = m_process->connect(gArgs.GetDataDirNet(), "adonai-node", address);
             } catch (const std::system_error& e) {
                 // If connection type is auto and socket path isn't accepting connections, or doesn't exist, catch the error and return null;
                 if (e.code() == std::errc::connection_refused || e.code() == std::errc::no_such_file_or_directory) {
@@ -78,7 +78,7 @@ public:
                 throw;
             }
         } else {
-            fd = m_process->connect(gArgs.GetDataDirNet(), "bitcoin-node", address);
+            fd = m_process->connect(gArgs.GetDataDirNet(), "adonai-node", address);
         }
         return m_protocol->connect(fd, m_exe_name);
     }
