+++ src/qt/test/CMakeLists.txt
@@ -4,29 +4,29 @@
 
 set(CMAKE_AUTOMOC_MOC_OPTIONS "-p${CMAKE_CURRENT_SOURCE_DIR}")
 
-add_executable(test_bitcoin-qt
+add_executable(test_adonai-qt
   apptests.cpp
   optiontests.cpp
   rpcnestedtests.cpp
   test_main.cpp
   uritests.cpp
   util.cpp
-  ../../init/bitcoin-qt.cpp
+  ../../init/adonai-qt.cpp
 )
 
-target_link_libraries(test_bitcoin-qt
+target_link_libraries(test_adonai-qt
   core_interface
-  bitcoinqt
+  adonaiqt
   test_util
-  bitcoin_node
+  adonai_node
   Boost::headers
   Qt6::Test
 )
 
-import_plugins(test_bitcoin-qt)
+import_plugins(test_adonai-qt)
 
 if(ENABLE_WALLET)
-  target_sources(test_bitcoin-qt
+  target_sources(test_adonai-qt
     PRIVATE
       addressbooktests.cpp
       wallettests.cpp
@@ -34,15 +34,15 @@ if(ENABLE_WALLET)
   )
 endif()
 
-add_test(NAME test_bitcoin-qt
-  COMMAND test_bitcoin-qt
+add_test(NAME test_adonai-qt
+  COMMAND test_adonai-qt
 )
 if(WIN32 AND VCPKG_TARGET_TRIPLET)
   # On Windows, vcpkg configures Qt with `-opengl dynamic`, which makes
   # the "minimal" platform plugin unusable due to internal Qt bugs.
-  set_tests_properties(test_bitcoin-qt PROPERTIES
+  set_tests_properties(test_adonai-qt PROPERTIES
     ENVIRONMENT "QT_QPA_PLATFORM=windows"
   )
 endif()
 
-install_binary_component(test_bitcoin-qt)
+install_binary_component(test_adonai-qt)
